use 64-bit indexers also in windows

+++ hpfrec/cython_loops.pyx
@@ -12,8 +12,8 @@ import ctypes
 ## parallel for loops with unsigend iterators. If you are using a different compiler, this part can be safely removed
 ## See also: https://github.com/cython/cython/issues/3136
 IF UNAME_SYSNAME == "Windows":
-	obj_ind_type = ctypes.c_long
-	ctypedef long ind_type
+	obj_ind_type = ctypes.c_longlong
+	ctypedef long long ind_type
 	ctypedef double long_double_type
 	obj_long_double_type = ctypes.c_double
 ELSE:
